requirements.txt
autoPyTorch==0.0.2
backcall==0.2.0
catboost==0.24.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
ConfigSpace==0.4.13
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
doctest-ignore-unicode==0.1.2
docutils==0.16
fasteners==0.15
future==0.18.2
graphviz==0.14.1
h2o==3.30.1.1
hpbandster==0.7.4
idna==2.10
imbalanced-learn==0.7.0
imblearn==0.0
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.16.0
kiwisolver==1.2.0
liac-arff==2.4.0
lightgbm==2.3.1
matplotlib==3.3.1
mock==4.0.2
monotonic==1.5
munch==2.5.0
netifaces==0.10.9
networkx==2.5
nose==1.3.7
numpy==1.19.1
openml==0.10.2
pandas==1.1.1
parso==0.7.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.9.0
pretrainedmodels==0.7.4
prettytable==0.7.2
prompt-toolkit==3.0.7
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.6.0
Pygments==2.6.1
pygraphviz==1.6
pynisher==0.5.0
pyparsing==2.4.7
Pyro4==4.80
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.2
serpent==1.30.2
six==1.15.0
statsmodels==0.12.0
tabulate==0.8.7
tensorboard-logger==0.1.0
threadpoolctl==2.1.0
torch==1.6.0
torchsummary==1.5.1
torchvision==0.7.0
tqdm==4.48.2
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
xmltodict==0.12.0
